refactor and add Bucket.remove_record
