+++ src/bucket.rs
@@ -3,6 +3,11 @@
 //    License, v. 2.0. If a copy of the MPL was not distributed with this
 //    file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+mod read;
+mod remove;
+mod update;
+mod write;
+
 use std::sync::Arc;
 
 use reqwest::Method;
@@ -13,12 +18,6 @@ use reduct_base::msg::entry_api::EntryInfo;
 
 use crate::client::Result;
 use crate::http_client::HttpClient;
-use crate::record::query::QueryBuilder;
-use crate::record::read_record::ReadRecordBuilder;
-use crate::record::update_record::UpdateRecordBuilder;
-use crate::record::write_batched_records::WriteBatchType;
-use crate::record::WriteRecordBuilder;
-use crate::WriteBatchBuilder;
 
 /// A bucket to store data in.
 pub struct Bucket {
@@ -169,207 +168,16 @@ impl Bucket {
     pub async fn entries(&self) -> Result<Vec<EntryInfo>> {
         Ok(self.full_info().await?.entries)
     }
-
-    /// Create a record to write to the bucket.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to write to.
-    ///
-    /// # Returns
-    ///
-    /// Returns a record builder.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use reduct_rs::{ReductClient, ReductError};
-    /// use std::time::SystemTime;
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), ReductError> {
-    ///    let client = ReductClient::builder()
-    ///         .url("https://play.reduct.store")
-    ///         .api_token("reductstore")
-    ///         .build();
-    ///     let bucket = client.get_bucket("demo").await?;
-    ///     let record = bucket.write_record("entry-1").timestamp_us(1000).data("Some data").send().await?;
-    ///     Ok(())
-    /// }
-    /// ```
-    pub fn write_record(&self, entry: &str) -> WriteRecordBuilder {
-        WriteRecordBuilder::new(
-            self.name.clone(),
-            entry.to_string(),
-            Arc::clone(&self.http_client),
-        )
-    }
-
-    /// Create a batch to write to the bucket.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to write to.
-    ///
-    /// # Returns
-    ///
-    /// Returns a batch builder.
-    pub fn write_batch(&self, entry: &str) -> WriteBatchBuilder {
-        WriteBatchBuilder::new(
-            self.name.clone(),
-            entry.to_string(),
-            Arc::clone(&self.http_client),
-            WriteBatchType::Write,
-        )
-    }
-
-    /// Update labels of a record in an entry.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to update.
-    ///
-    /// # Returns
-    ///
-    /// Returns a builder to update the record and send the request.
-    pub fn update_record(&self, entry: &str) -> UpdateRecordBuilder {
-        UpdateRecordBuilder::new(
-            self.name.clone(),
-            entry.to_string(),
-            Arc::clone(&self.http_client),
-        )
-    }
-
-    /// Create a batch to update records in the bucket.
-    ///
-    /// You should use RecordBuilder to create the records to update.
-    /// Add labels to the record to update. Labels with an empty value will be removed.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to update.
-    ///
-    /// # Returns
-    ///
-    /// Returns a batch builder.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use tokio;
-    /// use reduct_rs::{ReductClient, ReductError};
-    /// use reduct_rs::RecordBuilder;
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), ReductError> {
-    ///     let client = ReductClient::builder()
-    ///        .url("https://play.reduct.store")
-    ///        .api_token("reductstore")
-    ///        .build();
-    ///     let bucket = client.get_bucket("datasets").await?;
-    ///     let batch = bucket.update_batch("cats");
-    ///     let record1 = RecordBuilder::new()
-    ///         .timestamp_us(1000)
-    ///         .add_label("test".to_string(), "2".to_string())
-    ///         .add_label("x".to_string(), "".to_string()) // Remove label x
-    ///         .build();
-    ///
-    ///     batch.add_record(record1).send().await?;
-    ///     Ok(())
-    /// }
-    pub fn update_batch(&self, entry: &str) -> WriteBatchBuilder {
-        WriteBatchBuilder::new(
-            self.name.clone(),
-            entry.to_string(),
-            Arc::clone(&self.http_client),
-            WriteBatchType::Update,
-        )
-    }
-
-    /// Create a record to write to the bucket.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to write to.
-    ///
-    /// # Returns
-    ///
-    /// Returns a record builder.
-    pub fn read_record(&self, entry: &str) -> ReadRecordBuilder {
-        ReadRecordBuilder::new(
-            self.name.clone(),
-            entry.to_string(),
-            Arc::clone(&self.http_client),
-        )
-    }
-
-    /// Create a record to write to the bucket.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to write to.
-    ///
-    /// # Returns
-    ///
-    /// Returns a record builder.
-    ///
-    /// ```no_run
-    /// use reduct_rs::{ReductClient, ReductError};
-    /// use std::time::SystemTime;
-    /// use futures_util::StreamExt;
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> Result<(), ReductError> {
-    ///    let client = ReductClient::builder()
-    ///         .url("https://play.reduct.store")
-    ///         .api_token("reductstore")
-    ///         .build();
-    ///     let bucket = client.get_bucket("datasets").await?;
-    ///     let query = bucket.query("cats").limit(10).send().await?;
-    ///     tokio::pin!(query);
-    ///     while let Some(record) = query.next().await {
-    ///         let record = record?;
-    ///         let content_ = record.bytes().await?;
-    ///     }
-    ///     Ok(())
-    /// }
-    ///  ```
-    pub fn query(&self, entry: &str) -> QueryBuilder {
-        QueryBuilder::new(
-            self.name.clone(),
-            entry.to_string(),
-            Arc::clone(&self.http_client),
-        )
-    }
-
-    /// Remove an entry from the bucket.
-    ///
-    /// # Arguments
-    ///
-    /// * `entry` - The entry to remove.
-    ///
-    /// # Returns
-    ///
-    /// Returns an error if the entry could not be removed.
-    pub async fn remove_entry(&self, entry: &str) -> Result<()> {
-        let request = self
-            .http_client
-            .request(Method::DELETE, &format!("/b/{}/{}", self.name, entry));
-        self.http_client.send_request(request).await?;
-        Ok(())
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use bytes::Bytes;
-    use futures_util::{pin_mut, StreamExt};
     use rstest::{fixture, rstest};
 
     use reduct_base::error::ErrorCode;
 
     use crate::client::tests::{bucket_settings, client};
     use crate::client::ReductClient;
-    use crate::record::RecordBuilder;
 
     use super::*;
 
@@ -421,420 +229,8 @@ mod tests {
         );
     }
 
-    #[rstest]
-    #[tokio::test]
-    async fn test_bucket_write_record_data(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        bucket
-            .write_record("test")
-            .timestamp_us(1000)
-            .data(Bytes::from("Hey"))
-            .send()
-            .await
-            .unwrap();
-
-        let record = bucket
-            .read_record("test")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.bytes().await.unwrap(), Bytes::from("Hey"));
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_bucket_write_record_stream(#[future] bucket: Bucket) {
-        let chunks: Vec<Result<_>> = vec![Ok("hello"), Ok(" "), Ok("world")];
-
-        let stream = futures_util::stream::iter(chunks);
-
-        let bucket: Bucket = bucket.await;
-        bucket
-            .write_record("test")
-            .timestamp_us(1000)
-            .content_length(11)
-            .stream(Box::pin(stream))
-            .send()
-            .await
-            .unwrap();
-
-        let record = bucket
-            .read_record("test")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(record.bytes().await.unwrap(), Bytes::from("hello world"));
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_read_record(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        let record = bucket
-            .read_record("entry-1")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.timestamp_us(), 1000);
-        assert_eq!(record.content_length(), 12);
-        assert_eq!(record.content_type(), "text/plain");
-        assert_eq!(record.labels().get("bucket"), Some(&"1".to_string()));
-        assert_eq!(record.labels().get("entry"), Some(&"1".to_string()));
-        assert_eq!(record.bytes().await.unwrap(), Bytes::from("Hey entry-1!"));
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_read_record_as_stream(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        let record = bucket
-            .read_record("entry-1")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-
-        let mut stream = record.stream_bytes();
-        assert_eq!(
-            stream.next().await.unwrap(),
-            Ok(Bytes::from("Hey entry-1!"))
-        );
-        assert_eq!(stream.next().await, None);
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_head_record(#[future] bucket: Bucket) {
-        let record = bucket
-            .await
-            .read_record("entry-1")
-            .timestamp_us(1000)
-            .head_only(true)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.timestamp_us(), 1000);
-        assert_eq!(record.content_length(), 12);
-        assert_eq!(record.content_type(), "text/plain");
-        assert_eq!(record.labels().get("bucket"), Some(&"1".to_string()));
-        assert_eq!(record.labels().get("entry"), Some(&"1".to_string()));
-    }
-
-    mod query {
-        use super::*;
-        use chrono::Duration;
-        use std::time::SystemTime;
-
-        #[rstest]
-        #[case(true, 10)]
-        #[case(false, 100)]
-        #[case(false, 10_000)]
-        #[case(false, 20_000_000)]
-        #[tokio::test]
-        async fn test_query(
-            #[future] bucket: Bucket,
-            #[case] head_only: bool,
-            #[case] size: usize,
-        ) {
-            let bucket: Bucket = bucket.await;
-            let mut bodies: Vec<Vec<u8>> = Vec::new();
-            for i in 0..20usize {
-                let mut content = Vec::with_capacity(size);
-                for _j in 0..size {
-                    content.push(i as u8);
-                }
-                bodies.push(content);
-
-                bucket
-                    .write_record("entry-3")
-                    .timestamp_us(i as u64)
-                    .data(Bytes::from(bodies[i].clone()))
-                    .send()
-                    .await
-                    .unwrap();
-            }
-
-            let query = bucket
-                .query("entry-3")
-                .ttl(Duration::minutes(1).to_std().unwrap())
-                .head_only(head_only)
-                .send()
-                .await
-                .unwrap();
-            pin_mut!(query);
-
-            for i in 0..20usize {
-                let record = query.next().await.unwrap().unwrap();
-                assert_eq!(record.timestamp_us(), i as u64);
-                assert_eq!(record.content_length(), size);
-                assert_eq!(record.content_type(), "application/octet-stream");
-
-                if !head_only {
-                    assert_eq!(
-                        record.bytes().await.unwrap(),
-                        Bytes::from(bodies[i].clone())
-                    );
-                }
-            }
-
-            assert!(query.next().await.is_none());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_query_each_second(#[future] bucket: Bucket) {
-            let bucket: Bucket = bucket.await;
-            let query = bucket.query("entry-2").each_s(0.002).send().await.unwrap();
-
-            pin_mut!(query);
-            let rec = query.next().await.unwrap().unwrap();
-            assert_eq!(rec.timestamp_us(), 2000);
-            let rec = query.next().await.unwrap().unwrap();
-            assert_eq!(rec.timestamp_us(), 4000);
-            assert!(query.next().await.is_none());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_query_each_minute(#[future] bucket: Bucket) {
-            let bucket: Bucket = bucket.await;
-            let query = bucket.query("entry-2").each_n(2).send().await.unwrap();
-
-            pin_mut!(query);
-            let rec = query.next().await.unwrap().unwrap();
-            assert_eq!(rec.timestamp_us(), 2000);
-            let rec = query.next().await.unwrap().unwrap();
-            assert_eq!(rec.timestamp_us(), 4000);
-            assert!(query.next().await.is_none());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_limit_query(#[future] bucket: Bucket) {
-            let bucket: Bucket = bucket.await;
-            let query = bucket.query("entry-1").limit(1).send().await.unwrap();
-
-            pin_mut!(query);
-            let _ = query.next().await.unwrap().unwrap();
-            assert!(query.next().await.is_none());
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_query_with_few_options(#[future] bucket: Bucket) {
-            let bucket: Bucket = bucket.await;
-            let query = bucket
-                .query("entry-1")
-                .start(SystemTime::now() - Duration::minutes(1).to_std().unwrap())
-                .stop(SystemTime::now() + Duration::minutes(1).to_std().unwrap())
-                .limit(1)
-                .add_include("bucket", "1")
-                .add_exclude("entry", "1")
-                .send()
-                .await;
-            assert!(query.is_ok());
-        }
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn remove_entry(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        bucket.remove_entry("entry-1").await.unwrap();
-        assert!(
-            bucket
-                .read_record("entry-1")
-                .send()
-                .await
-                .err()
-                .unwrap()
-                .status
-                == ErrorCode::NotFound
-        );
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_batched_write(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        let batch = bucket.write_batch("test");
-
-        let record1 = RecordBuilder::new()
-            .timestamp_us(1000)
-            .data(Bytes::from("Hey,"))
-            .add_label("test", "1")
-            .content_type("text/plain")
-            .build();
-
-        let stream = futures_util::stream::iter(vec![Ok(Bytes::from("World"))]);
-
-        let record2 = RecordBuilder::new()
-            .timestamp_us(2000)
-            .stream(Box::pin(stream))
-            .add_label("test", "2")
-            .content_type("text/plain")
-            .content_length(5)
-            .build();
-
-        let error_map = batch
-            .add_record(record1)
-            .add_record(record2)
-            .send()
-            .await
-            .unwrap();
-        assert!(error_map.is_empty());
-
-        let record = bucket
-            .read_record("test")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.content_length(), 4);
-        assert_eq!(record.content_type(), "text/plain");
-        assert_eq!(record.labels().get("test"), Some(&"1".to_string()));
-        assert_eq!(record.bytes().await.unwrap(), Bytes::from("Hey,"));
-
-        let record = bucket
-            .read_record("test")
-            .timestamp_us(2000)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.content_length(), 5);
-        assert_eq!(record.content_type(), "text/plain");
-        assert_eq!(record.labels().get("test"), Some(&"2".to_string()));
-        assert_eq!(record.bytes().await.unwrap(), Bytes::from("World"));
-    }
-
-    #[rstest]
-    #[tokio::test]
-    async fn test_batched_write_with_error(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        bucket
-            .write_record("test")
-            .timestamp_us(1000)
-            .data(Bytes::from("xxx"))
-            .send()
-            .await
-            .unwrap();
-
-        let batch = bucket.write_batch("test");
-        let record1 = RecordBuilder::new()
-            .timestamp_us(1000)
-            .data(Bytes::from("Hey,"))
-            .build();
-        let record2 = RecordBuilder::new()
-            .timestamp_us(2000)
-            .data(Bytes::from("World"))
-            .build();
-
-        let error_map = batch
-            .add_record(record1)
-            .add_record(record2)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(error_map.len(), 1);
-        assert_eq!(error_map.get(&1000).unwrap().status, ErrorCode::Conflict);
-        assert_eq!(
-            error_map.get(&1000).unwrap().message,
-            "A record with timestamp 1000 already exists"
-        );
-    }
-
-    #[rstest]
-    #[tokio::test]
-    #[cfg_attr(not(feature = "test-api-111"), ignore)]
-    async fn test_update_record(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        bucket
-            .write_record("test")
-            .timestamp_us(1000)
-            .data(Bytes::from("Hey"))
-            .add_label("test", "1")
-            .add_label("x", "y")
-            .send()
-            .await
-            .unwrap();
-
-        bucket
-            .update_record("test")
-            .timestamp_us(1000)
-            .update_label("test", "2")
-            .remove_label("x")
-            .send()
-            .await
-            .unwrap();
-
-        let record = bucket
-            .read_record("test")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.labels().get("test"), Some(&"2".to_string()));
-    }
-
-    #[rstest]
-    #[tokio::test]
-    #[cfg_attr(not(feature = "test-api-111"), ignore)]
-    async fn test_update_record_batched(#[future] bucket: Bucket) {
-        let bucket: Bucket = bucket.await;
-        bucket
-            .write_record("test")
-            .timestamp_us(1000)
-            .data(Bytes::from("Hey"))
-            .add_label("test", "1")
-            .add_label("x", "y")
-            .send()
-            .await
-            .unwrap();
-
-        let batch = bucket.update_batch("test");
-        let record1 = RecordBuilder::new()
-            .timestamp_us(1000)
-            .add_label("test".to_string(), "2".to_string())
-            .add_label("x".to_string(), "".to_string())
-            .build();
-        let record2 = RecordBuilder::new()
-            .timestamp_us(10000)
-            .add_label("test".to_string(), "3".to_string())
-            .build();
-
-        let error_map = batch
-            .add_record(record1)
-            .add_record(record2)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(error_map.len(), 1);
-        assert_eq!(error_map.get(&10000).unwrap().status, ErrorCode::NotFound);
-
-        let record = bucket
-            .read_record("test")
-            .timestamp_us(1000)
-            .send()
-            .await
-            .unwrap();
-
-        assert_eq!(record.labels().get("test"), Some(&"2".to_string()));
-        assert_eq!(record.labels().get("x"), None);
-    }
-
     #[fixture]
-    async fn bucket(#[future] client: ReductClient) -> Bucket {
+    pub async fn bucket(#[future] client: ReductClient) -> Bucket {
         client.await.get_bucket("test-bucket-1").await.unwrap()
     }
 }
